Internals

Cell management

PlutoBoard.add_cell — Method

add_cell(
	;
	ws::WebSocket
) -> nothing

Adds a cell at the bottom of the notebook

PlutoBoard.remove_cell — Method

remove_cell(
	uuid::String
	;
	ws::WebSocket
) -> nothing

Removes cell with uuid

PlutoBoard.get_cells — Method

get_cells(
	;
	ws::WebSocket
) -> Array{String}

Returns an array of all cells uuids

PlutoBoard.find_cells_with_variable — Method

find_cells_with_variable(
    var::String
) -> Set{String}

Finds all cells in the notebook that contain the given variable and returns their IDs.

JS, HTML and CSS injections

PlutoBoard.load_scripts_and_links — Function

load_scripts_and_links() -> HypertextLiteral.Result

Returns a HypertextLiteral.Result object with scritps and links to load defined by the developer in initialize

PlutoBoard.load_js — Method

load_js() -> HypertextLiteral.Result

Returns a HypertextLiteral.Result object to load entry js files as modules and css stylesheets.

PlutoBoard.get_css_files — Method

get_css_files() -> Array{String}

Returns all css files in static/css folder of developer package.

Fileserver

PlutoBoard.serve_file — Method

serve_file(req::HTTP.Request)

Serves the requested file from the SERVE_DIR directory.

PlutoBoard.start_server — Method

start_server()

Starts the server on the port specified in the config.toml file.

PlutoBoard.run_fileserver — Method

run_fileserver()

Starts the file server with restarting if needed.

PlutoBoard.get_mime_type — Method

get_mime_type(file)

Returns the MIME type of the file.

PlutoBoard.monitor_folder — Method

monitor_folder(; ws)

Watches the folder specified by `SERVE_DIR` for changes. When a file change is detected, sends a notification to the provided WebSocket `ws`.

Other stuff

PlutoBoard.setup — Method

setup()::nothing

Sets up PlutoBoard project by creating a new Julia package and copying necessary files.

PlutoBoard.get_package_name — Method

get_package_name()::String

Returns the name of the package from the Project.toml file.

PlutoBoard.open_file — Method

open_file(
	path::String
)::String

Opens a file and returns its content as a string.

PlutoBoard.copy_with_delete — Method

copy_with_delete(
	from::String,
	to::String
)::nothing

Copies a file from one location to another and deletes the original file. Sets the permissions of the copied file to 666.

PlutoBoard.parse_to_symbol — Method

parse_to_symbol(
	d::Dict{String, Any}
)::Dict{Symbol, Any}

Parse a dictionary with string keys to a dictionary with symbol keys.