PlutoBoard.jl Documentation

PlutoBoard.jl Documentation

Interface functions

Internal functions

Interface functions

Julia interface

PlutoBoard.initialize — Method

initiliaize(
	html_path::String,
	css_path::String
	;
	fullscreen::Bool = false,
	bootstrap::Bool = false,
	hide_notebook::Bool = true
) -> nothing

Initializes the PlutoBoard module with parameters.

JavaScript interface

These functions can be called in the browser and are written in JavaScript.

PlutoBoard.callJuliaFunction — Method

async callJuliaFunction(
	func_name,
	{ args = [], kwargs = {}, response_callback = () => { } } = {}
) -> Promise

kwargs are kwargs for the Julia function to be called.

Calls specified Julia function with args and callback within browser context from JavaScript.

Example usage:

callJuliaFunction("get_cube", {
	args: [number],
	response_callback: (
		r => {
			const outP = document.getElementById("buttonOutput");
			outP.innerHTML = `Cube of ${number}... ${Math.round(r * 100)}%`;
		}
	)
	}).then(
		r => {
			const outP = document.getElementById("buttonOutput");
			outP.innerHTML = `The cube of ${number} is ${r}`;
		}
	)

Internal functions

Julia internal functions

PlutoBoard.load_bootstrap_css — Method

load_bootstrap_css() -> HypertextLiteral.Result

Returns a HypertextLiteral.Result object with the bootstrap css link.

PlutoBoard.load_bootstrap_js — Method

load_bootstrap_js() -> HypertextLiteral.Result

Returns a HypertextLiteral.Result object with the bootstrap js link.

PlutoBoard.load_html — Method

load_html() -> HTML

Returns a HTML object with the html and css loaded.

PlutoBoard.load_html_string_to_body — Method

load_html_string_to_body() -> HypertextLiteral.Result

Returns a HypertextLiteral.Result object to call a javascript function to insert the html and css to the body.

PlutoBoard.load_js — Method

load_js() -> HypertextLiteral.Result

Returns a HypertextLiteral.Result object to load all javascript files.

JavaScript internal functions

PlutoBoard.placeIframe — Method

placeIframe(targetCellID, destinationDiv)

Places an iframe of the site itself in destinationDiv only showing targetCellID.

PlutoBoard.updateAllCells — Method

async updateAllCells()

Forcefully reevaluates all Pluto cells.

PlutoBoard.updateCell — Method

async updateCell(cellid)

Reevaluates Pluto cell with cellid.

PlutoBoard.insertHTMLToBody — Method

insertHTMLToBody(html, css)

Appends html and css string below body.