Starting a new project

Setup

Write your own code

PlutoBoard.jl Documentation

PlutoBoard interface

Internal functions

Starting a new project

Setup

Generate your package

$> julia -e 'using Pkg; Pkg.generate("YOUR_PACKAGE_NAME")' && cd YOUR_PACKAGE_NAME

Add PlutoBoard and set it up

YOUR_PACKAGE_NAME$> julia --project -e 'using Pkg; Pkg.add(url="https://github.com/UniStuttgart-IKR/PlutoBoard.jl"); using PlutoBoard; PlutoBoard.setup()' 

Run the notebook

YOUR_PACKAGE_NAME$> julia --project -e 'using Pluto; Pluto.run(notebook="PlutoBoardNotebook.jl")'

Write your own code

There is some hierarchy:

main in src/Main.jl gets called in the beginning, so use this as julia entry point

functions that should be callable from js need to go into src/Functions.jl

all javascript files in static/javascript/ are getting executed, though static/javascript/javascript.js is the last one, so use this as js entry point

There is a simple example in src/Functions.jl, src/static/index.html and src/static/javascript/javascript.js about calling a julia function from javascript with callbacks.

Simple calling of a Julia function within HTML using a button and an input

First, lets write a simple Julia function in src/Functions.jl

function get_cube(num; ws)
    # iterate 50 times
	for i in 1:50
        # send each i to the frontend
		PlutoBoard.send_to_ws(ws, i/50)
        # sleep a little bit so we have time to see it
		sleep(0.05)
	end
    # finally return the cube of `num`
	return num^3
end

Now we create an input with id=buttonInput for our number in static/index.html

<input type="text" id="buttonInput" placeholder="Input number">

as well as a button

<button class="btn btn-light" id="calculateButton" onclick="calculateVeryHardStuff()">Calculate</button>

and a div to show our output

<div class="text-white" id="buttonOutput">Output: </div>

Since we used calculateVeryHardStuff in our button, we need to define it aswell. We do this in static/javascript/javascript.js

function calculateVeryHardStuff() {
    // get our input
    const input = document.getElementById("buttonInput");
    // get the value of input 
    const number = parseInt(input.value);

    // call our previously defined Julia function with arguments and callback
    callJuliaFunction(
        // "get_cube" is the function we want to call
        "get_cube", {
        // `get_cube` takes one argument, in this case we pass `number`
        args: [number],
        // also pass a callback when we get a response (remember, we are sending the progress of the for loop, i/50)
        response_callback: (
            // define the function
            r => {
                // get output div
                const outputP = document.getElementById("buttonOutput");
                // set innerHTML of output div to show the progress
                outputP.innerHTML = `${Math.round(r * 100)}%...`;
            }
        )
    })
    // after getting the return value we want to show it in the output div
    .then(
        r => {
            const outputP = document.getElementById("buttonOutput");
            outputP.innerHTML = `Cube of ${number} is ${r}`;
        }
    )
}

And that's it, we have created a simple application using HTML and JavaScript as input and Julia for calculating!

PlutoBoard.jl Documentation

PlutoBoard interface

Julia interface

PlutoBoard.initialize — Method

initiliaize(
	html_path::String,
	css_path::String
	;
	fullscreen::Bool = true,
	hide_notebook::Bool = true,
	scripts::Array{String} = [],
	links::Array{String} = [],
) -> nothing

Initializes the PlutoBoard module with parameters.

JavaScript interface

These functions can be called in the browser and are written in JavaScript.

PlutoBoard.callJuliaFunction — Method

async callJuliaFunction(
	func_name,
	{ args = [], kwargs = {}, response_callback = () => { } } = {}
) -> Promise

kwargs are kwargs for the Julia function to be called.

Calls specified Julia function with args and callback within browser context from JavaScript.

Example usage:

callJuliaFunction("get_cube", {
	args: [number],
	response_callback: (
		r => {
			const outP = document.getElementById("buttonOutput");
			outP.innerHTML = `Cube of ${number}... ${Math.round(r * 100)}%`;
		}
	)
	}).then(
		r => {
			const outP = document.getElementById("buttonOutput");
			outP.innerHTML = `The cube of ${number} is ${r}`;
		}
	)

Internal functions

Julia internal functions

PlutoBoard.load_scripts_and_links — Method

load_scripts_and_links() -> HypertextLiteral.Result

Returns a HypertextLiteral.Result object with scritps and links to load.

PlutoBoard.load_html — Method

load_html() -> HTML

Returns a HTML object with the html and css loaded.

PlutoBoard.load_html_string_to_body — Method

load_html_string_to_body() -> HypertextLiteral.Result

Returns a HypertextLiteral.Result object to call a javascript function to insert the html and css to the body.

PlutoBoard.load_js — Method

load_js() -> HypertextLiteral.Result

Returns a HypertextLiteral.Result object to load all javascript files.

PlutoBoard.open_file — Method

open_file(
	path::String
)::String

Opens a file and returns its content as a string.

PlutoBoard.copy_with_delete — Method

copy_with_delete(
	from::String,
	to::String
)::nothing

Copies a file from one location to another and deletes the original file. Sets the permissions of the copied file to 644.

PlutoBoard.setup — Method

setup(
	example::String="default"
)::nothing

Sets up the PlutoBoard module by copying the necessary files to the current working directory. Example can either be default or vue or vue_moving_cells.

PlutoBoard.get_package_name — Method

get_package_name()::String

Returns the name of the package from the Project.toml file.

JavaScript internal functions

PlutoBoard.placeIframe — Method

placeIframe(targetCellID, destinationDiv)

Places an iframe of the site itself in destinationDiv only showing targetCellID.

PlutoBoard.updateAllCells — Method

async updateAllCells()

Forcefully reevaluates all Pluto cells.

PlutoBoard.updateCell — Method

async updateCell(cellid)

Reevaluates Pluto cell with cellid.

PlutoBoard.insertHTMLToBody — Method

insertHTMLToBody(html, css)

Appends html and css string below body.